requirements.txt
fastapi
uvicorn
transformers
pytesseract
pillow
pdf2image
streamlit
torch
